test: fuzz channel transactions through compiled validators
