+++ .github/workflows/ci.yml
@@ -506,6 +506,10 @@ jobs:
         working-directory: cardano/offchain
         run: deno task test:channels
 
+      - name: Fuzz channel transactions with compiled validators
+        working-directory: cardano/offchain
+        run: deno task test:channels:fuzz
+
       - name: Test shutdown finalization
         working-directory: cardano/offchain
         run: deno task test:shutdown
