+++ cardano/offchain/deno.json
@@ -8,7 +8,8 @@
     "@lucid-evolution/provider": "npm:@lucid-evolution/provider@0.1.90",
     "@lucid-evolution/utils": "npm:@lucid-evolution/utils@0.1.66",
     "@noble/hashes/blake2b": "npm:@noble/hashes@1.8.0/blake2b",
-    "ws": "npm:ws@7.5.10"
+    "ws": "npm:ws@7.5.10",
+    "fast-check": "npm:fast-check@3.23.2"
   },
   "nodeModulesDir": "auto",
   "lint": {
@@ -20,9 +21,10 @@
     "clean": "rm -rf deployments",
     "fmt:check": "deno fmt --check",
     "lint": "deno lint",
-    "check": "deno check index.ts src/*.ts scripts/*.ts types/**/*.ts",
+    "check": "deno check index.ts src/*.ts src/testing/*.ts scripts/*.ts types/**/*.ts",
     "test:deployment": "deno test --allow-env --allow-read src/deployment.test.ts",
     "test:channels": "deno test --allow-env --allow-read src/channel-actions.test.ts",
+    "test:channels:fuzz": "deno test --allow-env --allow-read src/channel-actions.fuzz.test.ts",
     "test:shutdown": "deno test --allow-env --allow-read scripts/shutdown-deployment.test.ts",
     "shutdown:deployment": "deno run --env-file=.env.default --allow-net --allow-env --allow-read --allow-run --allow-ffi scripts/shutdown-deployment.ts",
     "start": "deno run --env-file=.env.default --allow-net --allow-env --allow-read --allow-run --allow-ffi --allow-write index.ts"
