+++ cardano/offchain/deno.lock
@@ -9,13 +9,14 @@
     "jsr:@std/internal@^1.0.12": "1.0.12",
     "jsr:@std/internal@^1.0.9": "1.0.12",
     "jsr:@std/path@^1.1.1": "1.1.2",
-    "npm:@lucid-evolution/lucid@0.4.29": "0.4.29_effect@3.19.2",
+    "npm:@lucid-evolution/lucid@0.4.29": "0.4.29_effect@3.19.2_fast-check@3.23.2",
     "npm:@lucid-evolution/provider@0.1.90": "0.1.90_effect@3.19.2",
     "npm:@lucid-evolution/utils@*": "0.1.66_effect@3.19.2_@harmoniclabs+plutus-data@1.2.6__@harmoniclabs+bytestring@1.0.0__@harmoniclabs+cbor@1.6.6",
     "npm:@lucid-evolution/utils@0.1.66": "0.1.66_effect@3.19.2_@harmoniclabs+plutus-data@1.2.6__@harmoniclabs+bytestring@1.0.0__@harmoniclabs+cbor@1.6.6",
     "npm:@noble/hashes@1.8.0": "1.8.0",
     "npm:@types/node@*": "24.2.0",
     "npm:cbor@10": "10.0.11",
+    "npm:fast-check@3.23.2": "3.23.2",
     "npm:ws@*": "7.5.10",
     "npm:ws@7.5.10": "7.5.10"
   },
@@ -302,7 +303,7 @@
     "@lucid-evolution/crc8@0.1.8": {
       "integrity": "sha512-88+/S6n+sPgxYtG05rvbJDVb8MoFGzckkR3p26uGrzngYzlJGfxBWIzNhVwfI2mgINPo2C5lkZW3GYnfY0MjuA=="
     },
-    "@lucid-evolution/lucid@0.4.29_effect@3.19.2": {
+    "@lucid-evolution/lucid@0.4.29_effect@3.19.2_fast-check@3.23.2": {
       "integrity": "sha512-+9pXlHivIGE4xJC6EZ62vtOBfgHgPpbJHgd7d6Ztk3lmbBwsM1PZVQ1WSE2upXSvqGE8KT/99IxmZ3kKfbtDUg==",
       "dependencies": [
         "@anastasia-labs/cardano-multiplatform-lib-browser@6.0.2-3",
@@ -1054,6 +1055,7 @@
       "npm:@lucid-evolution/provider@0.1.90",
       "npm:@lucid-evolution/utils@0.1.66",
       "npm:@noble/hashes@1.8.0",
+      "npm:fast-check@3.23.2",
       "npm:ws@7.5.10"
     ]
   }
