+++ cardano/onchain/validators/minting_connection_stt_contract.test.ak
@@ -445,7 +445,7 @@ fn run_conn_open_try_transition(
 }
 
 test prop_tx_conn_open_try_accepts_valid_full_transition(_seed via fuzz.int()) {
-  fuzz.label(@"regression.tx.conn.open_try.valid_full_transition")
+  fuzz.label(@"regression.contract.conn.open_try.valid_full_transition")
 
   // tx.* depth: the same ConnOpenTry transaction is checked by connection
   // minting, HostState CreateConnection, and the connection datum transition.
@@ -455,13 +455,15 @@ test prop_tx_conn_open_try_accepts_valid_full_transition(_seed via fuzz.int()) {
     fuzz_dsl.no_mutation(),
   )
     |> run_conn_open_try_transition
-    |> fuzz_dsl.assert_protocol_valid
+    |> fuzz_dsl.assert_protocol_checks_accept
 }
 
 test prop_conn_open_try_rejects_wrong_counterparty_client_state(
   _seed via fuzz.int(),
 ) {
-  fuzz.label(@"regression.tx.conn.open_try.invalid_wrong_client_state_proof")
+  fuzz.label(
+    @"regression.contract.conn.open_try.invalid_wrong_client_state_proof",
+  )
 
   // Near-valid ConnOpenTry: the local connection datum and proof plumbing are
   // valid, but the counterparty client state bytes differ from the proof value.
@@ -471,5 +473,5 @@ test prop_conn_open_try_rejects_wrong_counterparty_client_state(
     fuzz_dsl.mutation("wrong_counterparty_client_state"),
   )
     |> run_conn_open_try_transition
-    |> fuzz_dsl.assert_protocol_rejected
+    |> fuzz_dsl.assert_protocol_checks_reject
 }
