+++ scripts/ci/aiken-fuzz-required-labels.json
@@ -8,16 +8,15 @@
   "allowedDepthPrefixes": [
     "unit",
     "contract",
-    "tx",
     "model"
   ],
   "requiredLabels": [
-    "fuzz.contract.host.handle_packet.invalid_ack_capacity",
     "fuzz.contract.host.bind_port.invalid_unauthorized",
+    "fuzz.contract.host.handle_packet.invalid_ack_capacity",
     "fuzz.contract.host.handle_packet.invalid_channel_only_change",
-    "fuzz.contract.host.handle_packet.invalid_recv_channel_mutation",
     "fuzz.contract.host.handle_packet.invalid_commitment_capacity",
     "fuzz.contract.host.handle_packet.invalid_history_capacity",
+    "fuzz.contract.host.handle_packet.invalid_recv_channel_mutation",
     "fuzz.contract.host.handle_packet.invalid_wrong_packet_key",
     "fuzz.contract.host.handle_packet.valid_ack",
     "fuzz.contract.host.handle_packet.valid_recv",
@@ -42,8 +41,12 @@
     "fuzz.contract.trace.rollover.invalid_non_target_bucket_changed",
     "fuzz.contract.trace.rollover.invalid_old_shard_not_preserved",
     "fuzz.contract.trace.rollover.valid",
-    "fuzz.unit.host.bind_port.invalid_capacity",
-    "fuzz.unit.host.bind_port.invalid_identifier",
+    "fuzz.model.channel.open_close_preserves_state",
+    "fuzz.unit.chan.close.invalid_packet_sequence",
+    "fuzz.unit.chan.close_confirm.valid",
+    "fuzz.unit.chan.close_init.valid",
+    "fuzz.unit.chan.open_ack.valid",
+    "fuzz.unit.chan.open_confirm.valid",
     "fuzz.unit.client.frozen.rejects_update",
     "fuzz.unit.client.misbehaviour.invalid_monotonic_headers",
     "fuzz.unit.client.misbehaviour.invalid_same_header",
@@ -53,17 +56,26 @@
     "fuzz.unit.client.update.invalid_wrong_consensus_state",
     "fuzz.unit.client.update.valid_adjacent",
     "fuzz.unit.client.update.valid_non_adjacent",
+    "fuzz.unit.host.bind_port.invalid_capacity",
+    "fuzz.unit.host.bind_port.invalid_identifier",
     "fuzz.unit.packet.recv.invalid_history_capacity",
     "fuzz.unit.packet.send.invalid_commitment_capacity",
-    "regression.contract.chan.close_confirm.valid",
-    "regression.contract.chan.close_init.valid_or_rejected_by_app_policy",
-    "regression.contract.chan.open_ack.valid",
-    "regression.contract.chan.open_confirm.valid",
     "regression.contract.client.recovery.invalid_missing_authority",
     "regression.contract.client.recovery.valid_expired",
     "regression.contract.client.recovery.valid_frozen",
     "regression.contract.client.update.invalid_missing_host_state",
     "regression.contract.client.update.invalid_wrong_host_redeemer",
+    "regression.contract.conn.open_try.invalid_wrong_client_state_proof",
+    "regression.contract.conn.open_try.valid_full_transition",
+    "regression.contract.packet.recv.invalid_callback_acknowledgement",
+    "regression.contract.packet.recv.invalid_callback_packet_data",
+    "regression.contract.packet.recv.invalid_missing_application_callback",
+    "regression.contract.packet.recv.invalid_moved_port_binding",
+    "regression.contract.packet.recv.invalid_unrelated_successor_mutation",
+    "regression.contract.packet.recv.valid_atomic_successor_callback",
+    "regression.contract.packet.recv.valid_sink_mints_voucher",
+    "regression.contract.packet.send.invalid_missing_operation_marker",
+    "regression.contract.packet.send.valid_channel_host_marker",
     "regression.contract.proof.batch.invalid_extra_item",
     "regression.contract.proof.batch.invalid_item_swapped",
     "regression.contract.proof.batch.invalid_missing_item",
@@ -86,30 +98,17 @@
     "regression.contract.transfer.voucher_send.burns_exactly",
     "regression.contract.voucher.first_mint.invalid_wrong_description",
     "regression.contract.voucher.first_mint.invalid_wrong_full_denom",
+    "regression.contract.voucher.first_mint.invalid_wrong_metadata_registry_coupled",
     "regression.contract.voucher.first_mint.invalid_wrong_name",
     "regression.contract.voucher.first_mint.invalid_wrong_policy",
     "regression.contract.voucher.first_mint.invalid_wrong_reference_asset",
     "regression.contract.voucher.first_mint.invalid_wrong_ticker",
     "regression.contract.voucher.first_mint.valid_metadata",
-    "regression.model.channel.open_close_reject_send",
+    "regression.contract.voucher.first_mint.valid_registry_coupled",
     "regression.model.packet.send_ack.valid_sequence",
     "regression.model.packet.send_recv_ack",
     "regression.model.packet.send_timeout",
     "regression.model.trace.append_rollover_lookup",
-    "regression.tx.conn.open_try.invalid_wrong_client_state_proof",
-    "regression.tx.conn.open_try.valid_full_transition",
-    "regression.tx.packet.recv.invalid_callback_acknowledgement",
-    "regression.tx.packet.recv.invalid_callback_packet_data",
-    "regression.tx.packet.recv.invalid_missing_application_callback",
-    "regression.tx.packet.recv.invalid_moved_port_binding",
-    "regression.tx.packet.recv.invalid_unrelated_successor_mutation",
-    "regression.tx.packet.recv.valid_atomic_successor_callback",
-    "regression.tx.packet.recv.valid_sink_mints_voucher",
-    "regression.tx.packet.send.invalid_missing_operation_marker",
-    "regression.tx.packet.send.valid_channel_host_marker",
-    "regression.tx.voucher.first_mint.invalid_wrong_metadata_registry_coupled",
-    "regression.tx.voucher.first_mint.valid_registry_coupled",
-    "regression.unit.chan.closed.rejects_send",
     "regression.unit.client.misbehaviour.invalid_wrong_client_id",
     "regression.unit.conn.open_ack.invalid_wrong_proof_height",
     "regression.unit.conn.open_ack.valid",
